Move fetching layer from YOLO model function, add target_layer parameter to GradCAM
